backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
torch==2.2.0
transformers==4.41.2
huggingface_hub==0.23.2
opencv-python==4.9.0.80
scipy==1.13.0
httpx==0.27.0
pydantic==2.7.1
pytest==8.2.1
pytest-asyncio==0.23.7
python-dotenv==1.0.1
